requirements.txt
datasets 
transformers[torch]
streamlit
Pillow